Mount Harvard East Slopes
Added a partial route for Mount Harvard's east slopes up to about 12,300
feet. This is intended to be used as a winter route because there's a
trail that goes up to the east ridge prior to the end of this route; be
aware that this route does go under some steep and potentially suspect
slopes in winter.

Also added waypoints for the same route:
Frenchman Creek Trailhead, both 2WD and 4WD
A log bridge crossing of Frenchman Creek (this is an estimate and may be
corrected in the future; it's indended to give a general idea in winter,
not be exact)
Colorado Trail/Frenchman Creek Trail junction